+++ Src/FluentAssertions/AssertionExtensions.cs
@@ -205,9 +205,28 @@ namespace FluentAssertions
             return () => ForceEnumeration(enumerable);
         }
 
+        /// <summary>
+        /// Forces enumerating a collection of the provided <paramref name="subject"/>.
+        /// Should be used to assert that a method that uses the <c>yield</c> keyword throws a particular exception.
+        /// </summary>
+        /// <param name="subject">The object that exposes the method or property.</param>
+        /// <param name="enumerable">A reference to the method or property to force enumeration of.</param>
+        public static Action Enumerating<T, TResult>(this T subject, Func<T, IEnumerable<TResult>> enumerable)
+        {
+            return () => ForceEnumeration(subject, enumerable);
+        }
+
         private static void ForceEnumeration(Func<IEnumerable> enumerable)
         {
             foreach (object _ in enumerable())
+            {
+                // Do nothing
+            }
+        }
+
+        private static void ForceEnumeration<T>(T subject, Func<T, IEnumerable> enumerable)
+        {
+            foreach (object _ in enumerable(subject))
             {
                 // Do nothing
             }